feat: permList를 permMap으로 수정, article_content DTO를 삭제하고 articleDTO에 포함시킴
