+++ doc/examples.sql
@@ -44,7 +44,7 @@ CREATE TABLE `article` (
   `depth_level` INT UNSIGNED NOT NULL DEFAULT 0,
   `group_seq` INT UNSIGNED NOT NULL DEFAULT 0,
   `regdate` DATETIME NOT NULL,
-  `update` DATETIME,
+  `upddate` DATETIME,
   `category_id` INT UNSIGNED NOT NULL,
   `ip_address` VARCHAR(20) NOT NULL,
   `member_id` BIGINT(10) UNSIGNED NOT NULL,
@@ -61,7 +61,7 @@ CREATE TABLE `comment` (
   `depth_level` INT UNSIGNED NOT NULL DEFAULT 0,
   `group_seq` INT UNSIGNED NOT NULL DEFAULT 0,
   `regdate` DATETIME NOT NULL,
-  `update` DATETIME,
+  `upddate` DATETIME,
   `ip_address` VARCHAR(20) NOT NULL,
   `member_id` BIGINT(10) UNSIGNED NOT NULL,
   `is_deleted` BOOLEAN NOT NULL DEFAULT FALSE,
